Switch to load svg via data-uri
When connecting to existing puppeteer instance, it possibly wont have the file on the file system, if it's a remote instance.

By converting to data-uri we don't need to save a temp file, and can load it anywhere.

There doesn't seem to be a confirmed limit on max data uri size for chrome, however I've seen a couple of places claim 2mb